% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_idr.R
\name{calculate_idr}
\alias{calculate_idr}
\title{Perform IDR thresholding}
\usage{
calculate_idr(peak_file_1, peak_file_2, stringent = TRUE, out_dir)
}
\arguments{
\item{peak_file_1}{Path to the .narrowPeak file of biological replicate 1.}

\item{peak_file_2}{Path to the .narrowPeak file of biological replicate 2.}

\item{stringent}{if `TRUE` threshold at 0.01. If `FALSE`, threshold at 0.05.}

\item{out_dir}{Path to the directory where the high-confidence peaks are
written.}
}
\value{
A summary of the peaks that passed the IDR threshold and the path
to the .narrowPeak file containing the peaks that passed this threshold.
}
\description{
`calculate_idr()` performs IDR thresholding on a pair of peak sets.
High-confidence peaks that pass the threshold are written to a ".narrowPeak"
file in the specified `out_dir`
}
