% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservative_idr.R
\name{conservative_idr}
\alias{conservative_idr}
\title{Conservative IDR analysis}
\usage{
conservative_idr(
  rep_treat_1,
  rep_treat_2,
  rep_ctrl_1 = NULL,
  rep_ctrl_2 = NULL,
  out_dir,
  subdir_name = "conservative_analysis",
  ...
)
}
\arguments{
\item{rep_treat_1}{The path to the BAM file for biological replicate 1}

\item{rep_treat_2}{The path to the BAM file for biological replicate 2}

\item{out_dir}{A subdirectory containing the results from the analysis will
be created at the specified location.}

\item{...}{Additional parameters to be passed to the macs_call_peak
function.}

\item{paired_end}{Logical, indicating if the BAM file is paired-end.}
}
\value{
list...
}
\description{
\code{conservative_idr()} performs the conservative IDR analysis as defined
by ENCODE. The function writes output files to a subdirectory in the
specified \code{out_dir}.
}
\seealso{
\link[ConsensusPeak]{macs_call_peak()}
}
