% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{env_chipr}
\alias{env_chipr}
\title{Initialise an environment to run the python package ChIP-R.}
\format{
An object of class \code{BasiliskEnvironment} of length 1.
}
\usage{
env_chipr
}
\description{
Initialise an environment to run the python package ChIP-R.
}
\keyword{datasets}
