% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_counts_matrix.R
\name{feature_counts_matrix}
\alias{feature_counts_matrix}
\title{Generate a featureCounts matrix for a peak set}
\usage{
feature_counts_matrix(peak_file, bam_files)
}
\arguments{
\item{peak_file}{Path to the peak file.}

\item{bam_files}{Path to the BAM file(s) that the peak file was derived from.}
}
\description{
\code{feature_counts_matrix()} computes the number of reads from a BAM file
that fall within each peak of your peak set.
}
