% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bam.R
\name{filter_bam}
\alias{filter_bam}
\title{Filter BAM file to exclude the second mate in the read pair}
\usage{
filter_bam(bam_file, out_dir)
}
\arguments{
\item{bam_file}{Path to a paired-end BAM file.}

\item{out_dir}{Path to where the filtered BAM file will be written.}
}
\description{
\code{filter_bam()} filters a paired-end BAM file to exclude the second
mate in the read pair. The result is a single-ended BAM file.
}
