% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudoreplicates.R
\name{generate_pseudoreplicates}
\alias{generate_pseudoreplicates}
\title{Generate pseudoreplicate BAM files}
\usage{
generate_pseudoreplicates(
  pooled_bam,
  out_dir,
  is_paired = TRUE,
  is_control = FALSE
)
}
\arguments{
\item{pooled_bam}{The path to the pooled BAM file.}

\item{is_control}{Logical, indicating if the BAM file is a control.}

\item{output_dir}{The directory to save the output BAM files.}

\item{paired_end}{Logical, indicating if the BAM file is paired-end.}
}
\value{
A list containing paths to the pseudoreplicate BAM files.
}
\description{
Generate pseudoreplicate BAM files
}
