% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping_peaks.R
\name{histone}
\alias{histone}
\title{Call consensus peaks from histone data}
\usage{
histone(
  treat_files,
  control_files = NULL,
  out_dir,
  subdir_name = "histone_replicate_analysis",
  paired_end = FALSE,
  ...
)
}
\description{
\code{histone()} accepts multiple biological replicates and calls consensus
peaks. Use the findOverlapsOfPeaks() function from the ChIPpeakAnno package
}
