% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macs_call_peak.R
\name{macs_call_peak}
\alias{macs_call_peak}
\title{Call peaks using MACSr}
\usage{
macs_call_peak(..., name, out_dir = ".")
}
\arguments{
\item{...}{Arguments to be passed to \link[MACSr]{callpeak()}}

\item{out_dir}{Path to the output directory where all files will be written}

\item{out_name}{Prefix for the output file names}
}
\value{
Full paths to the created \code{narrowPeak} or \link{broadPeak}
file
}
\description{
This function is a wrapper of the \link[MACSr]{callpeak()} function from the
\code{MACSr} package.
}
\seealso{
\link[MACSr]{callpeak()} which this function wraps
}
