% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_replicates_mspc.R
\name{multiple_replicates_mspc}
\alias{multiple_replicates_mspc}
\title{Run Multiple Sample Peak Calling (MSPC)}
\usage{
multiple_replicates_mspc(peak_files, ...)
}
\arguments{
\item{peak_files}{A vector of peak files to be passed to \code{mspc()}}

\item{...}{Additional parameters to be passed to \code{mspc()}


This function requires dotnet 6.0.0 or higher as a system dependency
Advise the user to open a conda env and install dotnet here.}
}
\description{
This function is a wrapper of the \link{rmspc}{mspc()} function. The function
handles an arbitrary number of replicates. Best suited for n > 2.
}
