% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_idr.R
\name{optimal_idr}
\alias{optimal_idr}
\title{Optimal IDR analysis}
\usage{
optimal_idr(
  treat_files,
  control_files = NULL,
  is_paired,
  idr_stringent,
  out_dir,
  subdir_name = "optimal_idr_analysis",
  ...
)
}
\arguments{
\item{treat_files}{Character vector containing paths to the treatment BAM
files.}

\item{control_files}{Character vector containing paths to the control BAM
files.}

\item{is_paired}{Logical, specifying whether or not the BAM file is
paired-end.}

\item{idr_stringent}{Logical relating to the IDR threshold. If set to TRUE
then the threshold used by IDR is 0.01. If set to FALSE then the threshold
is 0.05. The default is TRUE.}

\item{out_dir}{Character specifying the name of the output directory in which
a subdirectory containing the output files will be created.}

\item{subdir_name}{Character specifying the name of the subdirectory that the
output files will be written to.}
}
\value{
Summary of the IDR output
}
\description{
\code{optimal_idr()} runs the optimal IDR analysis. This involves pooling of
the biological replicate BAM files, shuffling the reads splitting into
pseudoreplicates.
}
