% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_files.R
\name{pool_files}
\alias{pool_files}
\title{Pool Biological Replicates into a Single BAM File}
\usage{
pool_files(named_list, out_dir)
}
\arguments{
\item{named_list}{A named list containing paths to BAM files.
This list is outputted from the `prepare_named_list` function.

The list includes:
- "treatment_file_1" Path to the first treatment replicate.
- "treatment_file_2" Path to the second treatment replicate.
- "control_file_1" Optional. Path to the first control replicate.
- "control_file_2" Optional. Path to the second control replicate.}

\item{out_dir}{The directory where the merged BAM files will be saved.}
}
\value{
Vector containing file paths to the output BAM files
}
\description{
This function pools biological replicates into a single BAM file. It is
designed to work with the output from the `prepare_named_list` function,
which provides a named list of BAM file paths. If control files are provided,
they are also merged into a separate control BAM file.
}
\keyword{internal}
