% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_and_call.R
\name{prepare_and_call}
\alias{prepare_and_call}
\title{Prepare files for peak calling and call peaks}
\usage{
prepare_and_call(treat_files, control_files = NULL, out_dir, ...)
}
\arguments{
\item{...}{Additional parameters to pass to macs_call_peak

Prepare and call will be used in all analyses except optimal_idr. Editing this
function to accommodate pseudoreplicates adds unnecessary complexity. The optimal
idr function will use a custom version of this code}
}
\description{
\code{prepare_and_call} prepares the BAM files into a list and then passes
this list to the macs_call_peak() function. The ouput is a list containining
file paths to the resulting peak files.
}
