% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peak_file.R
\name{process_peak_file}
\alias{process_peak_file}
\title{Process peak files before input to MSPC.}
\usage{
process_peak_file(peak_list, out_dir)
}
\description{
Process peak files before input to MSPC.
}
\keyword{internal}
