% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chipr.R
\name{run_chipr}
\alias{run_chipr}
\title{Call consensus peaks using ChIP-R}
\usage{
run_chipr(
  peak_files,
  out_name = "chipr_out",
  minentries = 2,
  rankmethod = "pvalue",
  duphandling = "average",
  fragment = FALSE,
  seed = 0.5,
  alpha = 0.05,
  size = 20
)
}
\arguments{
\item{peak_files}{Character vector containing paths to the peak files in
either narrowPeak, broadPeak or regionPeak format.}

\item{out_name}{Prefix for the output file names.}

\item{minentries}{The minimum number of intersections a given peak must
satisfy.}

\item{rankmethod}{The ranking method used to rank peaks within
replicates. Options include "signalvalue", "pvalue" and "qvalue". The
default is "pvalue".}

\item{duphandling}{Method for breaking ties. The options are "average" or
"random". The default is "average".}

\item{fragment}{Logical specifying whether the input peaks should be subject
to high levels of fragmentation. The default is FALSE.}

\item{seed}{Numeric specifying a seed for when duphandling = "random".
Value must be between 0 and 1. The default is 0.5.}

\item{alpha}{Numeric cut-off value for deciding the set of reproducible
peaks. The default is 0.05.}

\item{size}{Numeric specifying the minimum peak size when peaks are
reconnected after fragmentation. The minimum peak size is ordinarily
determined by the size of surrounding peaks, but in the case that there are
none, this value will be used. The default is 20.}
}
\description{
\code{run_chipr()} is a wrapper for the Python package ChIP-R. ChIP-R
assesses the reproducibility of peaks across multiple biological replicates
by adapting the rank product statistic. Parameter descriptions and
defaults are taken from https://github.com/rhysnewell/ChIP-R.
}
