% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservative_idr.R
\name{conservative_idr}
\alias{conservative_idr}
\title{Conservative IDR analysis}
\usage{
conservative_idr(
  treat_files,
  control_files = NULL,
  out_dir,
  subdir_name = "conservative_idr_analysis",
  ...
)
}
\arguments{
\item{treat_files}{Character vector containing paths to the treatment BAM
files.}

\item{control_files}{Character vector containing paths to the control BAM
files.}

\item{out_dir}{Character specifying the name of the output directory in which
a subdirectory containing the output files will be created.}

\item{subdir_name}{Character specifying the name of the subdirectory that the
output files will be written to.}

\item{...}{Additional parameters to be passed to the \code{macs_call_peak()}
function.}
}
\value{
A list containing a summary of the IDR analysis along with the path
to the output files.
}
\description{
\code{conservative_idr()} performs the conservative IDR analysis as defined
by ENCODE. The function writes output files to a subdirectory in the
specified \code{out_dir}.
}
\seealso{
\link[ConsensusPeak]{macs_call_peak()}
}
