% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bam.R
\name{filter_bam}
\alias{filter_bam}
\title{Filter BAM file to exclude the second mate in the read pair}
\usage{
filter_bam(bam_file)
}
\description{
For Tn5-based methods like CUT&Tag and TIPseq, we recommend filtering
paired-end BAM files to exclude the second mate in the read pair.
}
