% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macs_call_peak.R
\name{macs_call_peak}
\alias{macs_call_peak}
\title{Call peaks using MACSr}
\usage{
macs_call_peak(..., out_name, out_dir = ".")
}
\arguments{
\item{...}{Arguments to be passed to MACSr::callpeak()}

\item{out_name}{Prefix for the output file names}

\item{out_dir}{Path to the output directory where all files will be written}
}
\description{
This function is a wrapper of the MACSr::callpeak() function, which is itself
a wrapper of the Python package MACS3.
}
\seealso{
[MACSr::callpeak()] which this function wraps
}
