% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_replicates_mspc.R
\name{multiple_replicates_mspc}
\alias{multiple_replicates_mspc}
\title{Call peaks with MACSr and generate a consensus set with Multiple Sample Peak
Calling (MSPC)}
\usage{
multiple_replicates_mspc(
  treat_files,
  control_files = NULL,
  is_paired,
  out_dir,
  subdir_name = "mspc_analysis",
  replicateType,
  stringencyThreshold,
  weakThreshold,
  gamma = stringencyThreshold,
  c,
  alpha = 0.05,
  keep = TRUE,
  ...
)
}
\arguments{
\item{treat_files}{Character vector containing paths to the treatment BAM
files.}

\item{control_files}{Character vector containing paths to the control BAM
files.}

\item{is_paired}{Logical, specifying whether or not the BAM file is
paired-end.}

\item{out_dir}{Character specifying the name of the output directory in which
a subdirectory containing the output files will be created.}

\item{subdir_name}{Character specifying the name of the subdirectory that the
output files will be written to.}

\item{replicateType}{Character string. This argument defines
the replicate type. Possible values of the argument : 
'Bio','Biological', 'Tec', 'Technical'. 
This arguments is required . More information in Details.}

\item{stringencyThreshold}{Double. A threshold on p-values,
where peaks with p-value lower than this threshold, are
considered stringent.This arguments is required}

\item{weakThreshold}{Double. A threshold on p-values,
such that peaks with p-value between this and stringency
threshold, are considered weak peaks. This arguments is required}

\item{gamma}{Double. The combined stringency threshold.
Peaks with combined p-value below this threshold are confirmed.
Default value is the value of the argument stringencyThreshold.
This arguments is optional.}

\item{c}{Integer or character string. the minimum number of overlapping peaks
required before the MSPC program combines their p-value.
The value of c can be given in absolute (e.g., c = 2 will require
at least 2 samples) or percentage of input samples (e.g., c = 50% 
will require at least 50% of input samples) formats.
Default value is 1. This arguments is optional.
More information in Details.}

\item{alpha}{Double. The threshold for Benjamini-Hochberg
multiple testing correction.
Default value is 0.05. This arguments is optional.}

\item{keep}{Logical. This argument determines whether the mspc function
should keep or delete all the files generated while running
the MSPC program.This is an optional argument.
More information in Details.}
}
\value{
A list containing a summary of the MSPC along with the path to the
output files.
}
\description{
\code{multiple_replicates_mspc()} is a wrapper of the \link{rmspc}{mspc()}
function.
}
\examples{
\dontrun{
input1 <- testthat::test_path("testdata", "r1_test_creb.bam")
input2 <- testthat::test_path("testdata", "r2_test_creb.bam")
input3 <- testthat::test_path("testdata", "r3_test_creb.bam")

multiple_replicates_mspc(treat_files = c(input1, input2, input3),
                         out_dir = tempdir(),
                         subdir_name = "mspc_analysis"
                         is_paired = FALSE,
                         replicateType = "Biological",
                         stringencyThreshold = 1e-8,
                         weakThreshold = 1e-4,
                         c = 3,
                         ...
                         )
                         }

}
