% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_idr.R
\name{optimal_idr}
\alias{optimal_idr}
\title{Optimal IDR analysis}
\usage{
optimal_idr(
  treat_files,
  control_files = NULL,
  out_dir,
  subdir_name = "optimal_idr_analysis",
  paired_end = FALSE,
  ...
)
}
\arguments{
\item{treat_files}{Character vector containing paths to the treatment BAM
files.}

\item{control_files}{Character vector containing paths to the control BAM
files.}

\item{out_dir}{Character specifying the name of the output directory in which
a subdirectory containing the output files will be created.}

\item{subdir_name}{Character specifying the name of the subdirectory that the
output files will be written to.}

\item{paired_end}{Logical, indicating if the BAM file is paired-end. This
controls the behaviour of pseudoreplicate generation. Default is
\code{FALSE}.}

\item{...}{Additional parameters to be passed to the \code{macs_call_peak()}
function.}
}
\value{
Summary of the IDR output
}
\description{
\code{optimal(idr)} runs the optimal IDR analysis. This involves pooling of
the biological replicate BAM files, shuffling the reads and  and splitting
into pseudoreplicates.
}
